fix titles in FastAPI_generated OpenAPI (GDEV-1406)
When custom exceptions based on HttpCustomExceptionBase where used
together with FastAPI to generate an OpenAPI documentation,
the titles were wrong.

The titles were attempted to be set explicitly in the base class.
That worked fine for the pydantic-based schema validation, however caused problems wit
the OpenAPI generator of FastAPI,
which automatically sets the title based
on the class name.
Simply removing the explicit title definition thus resolved the problem.

Bumps version to 0.2.4.

Co-authored-by: Moritz Hahn <75744178+MoritzHahn1337@users.noreply.github.com>
Co-authored-by: Thomas Zajac <thomas-jakob.zajac@uni-tuebingen.de>